Ignore error 2150858843 during command cleanup
It may happen that winrm shell disappear right after the execution of a
command (before its cleanup).

It can happen especially if winrm is stopped or restarted following the
execution of the command. For instance, using test-kitchen with ec2
boxes will trigger some winrm daemon configuration changes (creation
phase) which will lead to a restart of winrm process.

In this cases, we don't really care to cleanup correctly the command
environment (winrm shell is already missing, no need to destroy it)

Fixes #253

Change-Id: Ic2583f7e1a9df9b3a7fe1a33517209de97a96923
Signed-off-by: Grégoire Seux <g.seux@criteo.com>