Do full enumeration of run_wql requests
Implement WSENUM pulls in order to get all responses from a wql query.
Allow a block to be passed to run_wql so that items can be processed as they're queried instead of building a large Array first. This will reduce memory usage and total run time for some large queries