Create thread in shell finalizer
The following call chain was observed:
1. Cmd.close_shell or Powershell.close_shell
2. HttpTransport#send_request
3. HTTPClient#post

And HTTPClient#post internally calls Mutex#synchronize, which raises
ThreadError if called from a finalizer. Although the behavior is not
documented, it makes sense as a finalizer can be called in a block
passed to Mutex#synchronize, and calling Mutex#syncrhonize again
for the same mutex in the finalizer can lead to deadlock.

Create a thread in the shell finalizer and do everything in the thread
to avoid problems with Mutex#synchronize.